fixing migration and model issues

+++ db/migrate/20220916091923_create_items.rb
@@ -2,7 +2,7 @@ class CreateItems < ActiveRecord::Migration[6.1]
   def change
     create_table :items do |t|
       t.string :name
-      t.double :price
+      t.decimal :price
       t.integer :quantity
       t.integer :sub_categories_id
       t.integer :categories_id
